Extract the RFC7940 test case to its own module
